Fix update route fix deleting own user login details and logout from sidebar
